requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.1
lxml>=4.9.1
difflib>=0.1
openai>=1.0.0
python-dotenv>=0.20.0
htmldiff>=0.1
tenacity>=8.1.0
validators>=0.20.0
colorama>=0.4.6
tqdm>=4.64.1
